requirements.txt
# ========================================
# dLLM Reasoning 依赖列表
# 基于 dllm_zihan 环境 (测试通过)
# ========================================

# Python 版本要求: >=3.10

# 核心深度学习框架
torch==2.5.1
transformers==4.57.1

# 配置管理
hydra-core==1.3.2

# 张量操作
tensordict==0.10.0

# 数据处理
pandas==2.3.3
pyarrow==22.0.0

# 训练监控
wandb==0.22.3

# ========================================
# VERL 框架 (必需)
# ========================================
# VERL 需要从源码安装，不在 PyPI 上
# 安装方法见项目 README 或:
# git clone https://github.com/volcengine/verl
# cd verl && pip install -e .
